refactor: Comment 컴포넌트 세분화
