+++ src/features/comment-add/ui/CommentAddModal.tsx
@@ -1,0 +1,24 @@
+import { useCommentContext } from "../../../shared/model/CommentContext"
+import { Button, Dialog, DialogContent, DialogHeader, DialogTitle, Textarea } from "../../../shared/ui"
+
+export const CommentAddModal = () => {
+  const { showAddCommentDialog, setShowAddCommentDialog, newComment, setNewComment, addComment } = useCommentContext()
+
+  return (
+    <Dialog open={showAddCommentDialog} onOpenChange={setShowAddCommentDialog}>
+      <DialogContent>
+        <DialogHeader>
+          <DialogTitle>새 댓글 추가</DialogTitle>
+        </DialogHeader>
+        <div className="space-y-4">
+          <Textarea
+            placeholder="댓글 내용"
+            value={newComment.body}
+            onChange={(e) => setNewComment({ ...newComment, body: e.target.value })}
+          />
+          <Button onClick={addComment}>댓글 추가</Button>
+        </div>
+      </DialogContent>
+    </Dialog>
+  )
+}
