+++ src/features/comment-item/ui/CommentItem.tsx
@@ -1,0 +1,50 @@
+import { Edit2, ThumbsUp, Trash2 } from "lucide-react"
+import { Button } from "../../../shared/ui"
+import { highlightText } from "../../../shared/lib/highlightText"
+import { useCommentContext } from "../../../shared/model/CommentContext"
+import { Comment } from "../../../entities/comment/types"
+import { usePostParamsContext } from "../../../shared/model/PostParamsContext"
+
+interface Props {
+  postId: number
+  comment: Comment
+}
+
+export const CommentItem = ({ postId, comment }: Props) => {
+  const {
+    likeComment,
+    deleteComment,
+
+    setSelectedComment,
+    setShowEditCommentDialog,
+  } = useCommentContext()
+  const { searchQuery } = usePostParamsContext()
+
+  return (
+    <div key={comment.id} className="flex items-center justify-between text-sm border-b pb-1">
+      <div className="flex items-center space-x-2 overflow-hidden">
+        <span className="font-medium truncate">{comment.user.username}:</span>
+        <span className="truncate">{highlightText(comment.body, searchQuery)}</span>
+      </div>
+      <div className="flex items-center space-x-1">
+        <Button variant="ghost" size="sm" onClick={() => likeComment(comment.id, postId)}>
+          <ThumbsUp className="w-3 h-3" />
+          <span className="ml-1 text-xs">{comment.likes}</span>
+        </Button>
+        <Button
+          variant="ghost"
+          size="sm"
+          onClick={() => {
+            setSelectedComment(comment)
+            setShowEditCommentDialog(true)
+          }}
+        >
+          <Edit2 className="w-3 h-3" />
+        </Button>
+        <Button variant="ghost" size="sm" onClick={() => deleteComment(comment.id, postId)}>
+          <Trash2 className="w-3 h-3" />
+        </Button>
+      </div>
+    </div>
+  )
+}
