+++ src/features/post-detail/ui/PostDetailModal.tsx
@@ -1,6 +1,6 @@
 import { Dialog, DialogContent, DialogHeader, DialogTitle } from "../../../shared/ui"
 import { highlightText } from "../../../shared/lib/highlightText"
-import { Comments } from "../../comment/ui/Comments"
+import { CommentView } from "../../../widgets/comment/ui/CommentView"
 import { usePostsContext } from "../../../shared/model/PostContext"
 import { usePostParamsContext } from "../../../shared/model/PostParamsContext"
 
@@ -16,7 +16,7 @@ export const PostDetailModal = () => {
         </DialogHeader>
         <div className="space-y-4">
           <p>{highlightText(selectedPost?.body, searchQuery)}</p>
-          {selectedPost && <Comments postId={selectedPost.id} searchQuery={searchQuery} />}
+          {selectedPost && <CommentView postId={selectedPost.id} />}
         </div>
       </DialogContent>
     </Dialog>
